requirements.txt
Django==1.10
future==0.15.2
oauthlib==2.0.0
psycopg2==2.6.2
python-twitter==3.1
requests==2.11.1
requests-oauthlib==0.6.2
redis==2.10.5
django-redis-cache==1.7.0